+++ src/main.rs
@@ -35,6 +35,9 @@ fn main() -> Result<()> {
             SelfCommand::Update(update_args) => {
                 commands::self_update::run(update_args.clone(), &config, &cli)?
             }
+            SelfCommand::Completions(comp_args) => {
+                commands::completions::run(comp_args.clone())?
+            }
         },
     };